dags/parallel_maintenance_poc.py
"""
Parallel Maintenance PoC DAG for IndexAgent

This is a proof-of-concept DAG demonstrating the parallel processing architecture
designed for IndexAgent. It showcases Dynamic Task Mapping patterns that would be
used with Git worktrees and Claude Code invocations.

PRODUCTION vs POC:
- POC: Uses mock shard data and simulated processing
- PRODUCTION: Would discover real files and invoke Claude CLI
- POC: Uses dict for status tracking
- PRODUCTION: Would use SQLite database at indexagent_meta/shard_status.db
- POC: Logs worktree operations
- PRODUCTION: Would create actual Git worktrees
"""

from datetime import datetime, timedelta
import random
import time
from typing import Dict, List, Any
import logging

from airflow import DAG
from airflow.decorators import task
from airflow.models import Variable
from datetime import datetime, timezone

# Configuration
PARALLEL_LIMIT = int(Variable.get("INDEXAGENT_PARALLEL_LIMIT", default_var=10))
DEFAULT_RETRY_DELAY = timedelta(minutes=5)
DEFAULT_RETRIES = 2

# Mock data for PoC
MOCK_SHARDS = [
    {"shard_id": "todo-001", "file_path": "src/example.py", "type": "todo", "priority": 1},
    {"shard_id": "todo-002", "file_path": "src/models.py", "type": "todo", "priority": 2},
    {"shard_id": "todo-003", "file_path": "src/utils.py", "type": "todo", "priority": 1},
    {"shard_id": "doc-001", "file_path": "src/parallel/worktree_manager.py", "type": "doc", "priority": 3},
    {"shard_id": "doc-002", "file_path": "src/parallel/claude_runner.py", "type": "doc", "priority": 3},
    {"shard_id": "doc-003", "file_path": "src/parallel/shard_coordinator.py", "type": "doc", "priority": 2},
    {"shard_id": "test-001", "file_path": "tests/test_example.py", "type": "test", "priority": 1},
    {"shard_id": "test-002", "file_path": "tests/test_models.py", "type": "test", "priority": 2},
]

# Default DAG arguments
default_args = {
    'owner': 'indexagent',
    'depends_on_past': False,
    'start_date': datetime.now(timezone.utc) - timedelta(days=1),
    'email_on_failure': False,
    'email_on_retry': False,
    'retries': DEFAULT_RETRIES,
    'retry_delay': DEFAULT_RETRY_DELAY,
}

# Create the DAG
dag = DAG(
    'parallel_maintenance_poc',
    default_args=default_args,
    description='PoC DAG demonstrating parallel processing with Dynamic Task Mapping',
    schedule=None,  # Manual trigger only for PoC
    catchup=False,
    max_active_tasks=PARALLEL_LIMIT,  # Limit total parallel tasks
    tags=['poc', 'parallel', 'indexagent'],
)

@task(dag=dag)
def discover_shards() -> List[Dict[str, Any]]:
    """
    Discover work units (shards) to process.
    
    POC: Returns mock shard data
    PRODUCTION: Would scan repository for files needing maintenance
    """
    logger = logging.getLogger(__name__)
    logger.info(f"Starting shard discovery with PARALLEL_LIMIT={PARALLEL_LIMIT}")
    
    # In production, this would:
    # 1. Connect to ShardCoordinator
    # 2. Scan repository for files with TODOs, missing docs, etc.
    # 3. Check shard_status.db for already processed shards
    # 4. Return list of shards needing processing
    
    # For PoC, return mock shards
    discovered_shards = MOCK_SHARDS.copy()
    logger.info(f"Discovered {len(discovered_shards)} shards for processing")
    
    # Sort by priority (in production, might use more complex logic)
    discovered_shards.sort(key=lambda x: x['priority'])
    
    return discovered_shards

@task(dag=dag, max_active_tis_per_dag=PARALLEL_LIMIT)
def process_shard(shard: Dict[str, Any]) -> Dict[str, Any]:
    """
    Process a single shard (file) in parallel.
    
    POC: Simulates processing with sleep and random success/failure
    PRODUCTION: Would create worktree, invoke Claude CLI, commit changes
    """
    logger = logging.getLogger(__name__)
    shard_id = shard['shard_id']
    file_path = shard['file_path']
    shard_type = shard['type']
    
    logger.info(f"Processing shard {shard_id}: {file_path} (type: {shard_type})")
    
    # Simulate worktree creation
    worktree_path = f"/tmp/worktrees/{shard_id}"
    branch_name = f"cc-parallel/{shard_id}"
    logger.info(f"[MOCK] Creating worktree at {worktree_path} on branch {branch_name}")
    
    # In production, this would:
    # 1. Use WorktreeManager to create Git worktree
    # 2. Checkout to branch cc-parallel/<shard-id>
    # 3. Prepare Claude context based on shard type
    
    # Simulate Claude invocation
    processing_time = random.uniform(2, 5)  # Simulate variable processing time
    logger.info(f"[MOCK] Invoking Claude for {shard_type} maintenance on {file_path}")
    time.sleep(processing_time)
    
    # In production, this would:
    # 1. Use ClaudeRunner to invoke Claude CLI
    # 2. Pass appropriate prompts based on shard type
    # 3. Apply Claude's suggested changes
    # 4. Commit changes to branch
    
    # Simulate random success/failure (90% success rate)
    success = random.random() > 0.1
    
    if success:
        logger.info(f"[MOCK] Successfully processed shard {shard_id}")
        # Simulate status update
        logger.info(f"[MOCK] Updating shard_status.db: {shard_id} -> COMPLETED")
        
        # In production, would update SQLite database
        result = {
            'shard_id': shard_id,
            'status': 'completed',
            'file_path': file_path,
            'type': shard_type,
            'branch': branch_name,
            'changes_made': random.randint(1, 5),  # Mock number of changes
            'processing_time': processing_time,
        }
    else:
        logger.error(f"[MOCK] Failed to process shard {shard_id}")
        # Simulate status update
        logger.info(f"[MOCK] Updating shard_status.db: {shard_id} -> FAILED")
        
        result = {
            'shard_id': shard_id,
            'status': 'failed',
            'file_path': file_path,
            'type': shard_type,
            'error': 'Simulated Claude invocation failure',
        }
        
        # In production, would raise exception to trigger retry
        # For PoC, we'll continue to show the pattern
    
    # Simulate worktree cleanup
    logger.info(f"[MOCK] Removing worktree at {worktree_path}")
    
    # In production, this would:
    # 1. Use WorktreeManager to remove worktree
    # 2. Clean up any temporary files
    
    return result

@task(dag=dag)
def aggregate_results(results: List[Dict[str, Any]]) -> Dict[str, Any]:
    """
    Aggregate results from all processed shards.
    
    POC: Summarizes mock processing results
    PRODUCTION: Would create PR, update tracking, send notifications
    """
    logger = logging.getLogger(__name__)
    logger.info(f"Aggregating results from {len(results)} processed shards")
    
    # Analyze results
    completed = [r for r in results if r.get('status') == 'completed']
    failed = [r for r in results if r.get('status') == 'failed']
    
    total_changes = sum(r.get('changes_made', 0) for r in completed)
    total_time = sum(r.get('processing_time', 0) for r in results)
    
    # Group by type
    by_type = {}
    for result in results:
        shard_type = result.get('type', 'unknown')
        if shard_type not in by_type:
            by_type[shard_type] = {'completed': 0, 'failed': 0}
        
        if result.get('status') == 'completed':
            by_type[shard_type]['completed'] += 1
        else:
            by_type[shard_type]['failed'] += 1
    
    summary = {
        'total_shards': len(results),
        'completed': len(completed),
        'failed': len(failed),
        'total_changes': total_changes,
        'total_processing_time': round(total_time, 2),
        'by_type': by_type,
        'branches_created': [r.get('branch') for r in completed if r.get('branch')],
    }
    
    logger.info(f"Processing summary: {summary}")
    
    # In production, this would:
    # 1. Create a pull request with all branches
    # 2. Update project metrics/dashboards
    # 3. Send notifications to relevant channels
    # 4. Clean up any remaining resources
    
    if failed:
        logger.warning(f"Failed shards: {[f['shard_id'] for f in failed]}")
        # In production, might trigger alerts or manual review
    
    return summary

# Define the DAG structure
with dag:
    # Discover shards to process
    shards = discover_shards()
    
    # Process each shard in parallel using Dynamic Task Mapping
    # The expand() method creates one task instance per shard
    processed = process_shard.expand(shard=shards)
    
    # Aggregate all results
    summary = aggregate_results(processed)

# Additional notes for production implementation:
"""
PRODUCTION ENHANCEMENTS:

1. ShardCoordinator Integration:
   - Connect to actual file discovery logic
   - Implement priority queue for shard assignment
   - Add shard locking to prevent duplicate processing

2. WorktreeManager Integration:
   - Create actual Git worktrees
   - Handle worktree lifecycle (create, use, cleanup)
   - Implement proper error handling for Git operations

3. ClaudeRunner Integration:
   - Invoke actual Claude CLI commands
   - Handle Claude API rate limits
   - Implement prompt templates for different maintenance types

4. Status Database:
   - Use SQLite at indexagent_meta/shard_status.db
   - Track detailed processing history
   - Enable resumption of interrupted runs

5. Error Handling:
   - Implement proper exception hierarchy
   - Add detailed logging for debugging
   - Create alerts for critical failures

6. Monitoring:
   - Add Airflow metrics for processing times
   - Track Claude API usage
   - Monitor resource utilization

7. Configuration:
   - Move settings to Airflow Variables/Connections
   - Add environment-specific configurations
   - Implement feature flags for gradual rollout
"""